refactor: unify function and request context
Add a Cotnext base class
fix: circular references between BaseAssistant, FunctionContext
